北北基公車搜尋改善&ios8 閃退修正

+++ NTOU_Mobile/module/ShuttleTrack/NTOUBusTableViewController.m
@@ -51,7 +51,7 @@
 
 #pragma mark - Table view data source
 
--(float)tableView:(UITableView *)tableView heightForHeaderInSection:(NSInteger)section
+-(CGFloat)tableView:(UITableView *)tableView heightForHeaderInSection:(NSInteger)section
 {
     return 32;
 }
